back/pyproject.toml
[tool.poetry]
name = "vacancy"
version = "0.1.0"
description = ""
authors = ["Andrey Petritsa <andrey@petritsa.ru>"]

[[tool.poetry.packages]]
include = "vacancy"
from = "vacancy_back"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
flask = ">=2.3,<3.0"
requests = ">=2.32.5,<3.0.0"
pyrogram = ">=2.0.106,<3.0.0"
tgcrypto = "*"
sqlalchemy = ">=2.0.43,<3.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
loguru = "^0.7.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.test.dependencies]
pytest = ">=8.0.0,<9.0.0"
behave = ">=1.3.3,<2.0.0"
